Сделан ProfileViewController
