+++ Navigation/Navigation/Extensions.swift
@@ -67,22 +67,6 @@ extension LogInViewController {
     }
 }
 
-// расширение Delegate и DataSource для ProfileViewController
-extension ProfileViewController: UITableViewDelegate,UITableViewDataSource {
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCell(withIdentifier: "reuseIdentifier", for: indexPath)
-        cell.textLabel!.text = "it works"
-        return cell
-    }
-
-    func numberOfSectionsInTableView(tableView: UITableView) -> Int {
-        return 0
-    }
-
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return 0
-    }
-}
 
 extension UIView {
     static var identifier: String {
