+++ Navigation/Navigation/Profile/ProfileViewController.swift
@@ -7,40 +7,109 @@
 
 import UIKit
 
-protocol DetailDelegate: AnyObject {
-    func changeText(_ text: String, indexPath: IndexPath)
-}
-
-class ProfileViewController: UIViewController {
+final class ProfileViewController: UIViewController {
     
-    private let profileHeaderView = ProfileHeaderView()
+    private var postModel: [Post] = Post.makePost()
     
-    
-    // добавляем элементы на контроллер
-    func setupProfileViewController() {
-        profileHeaderView.translatesAutoresizingMaskIntoConstraints = false
-        view.addSubview(profileHeaderView)
-        
-        
-        NSLayoutConstraint.activate([
-            profileHeaderView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor),
-            profileHeaderView.leadingAnchor.constraint(equalTo: view.leadingAnchor),
-            profileHeaderView.trailingAnchor.constraint(equalTo: view.trailingAnchor),
-            profileHeaderView.heightAnchor.constraint(equalToConstant: 220)
-        ])
-    }
-    
+    private lazy var tableView: UITableView = {
+        let tableView = UITableView(frame: .zero, style: .plain)
+        tableView.translatesAutoresizingMaskIntoConstraints = false
+        tableView.dataSource = self
+        tableView.delegate = self
+        tableView.rowHeight = UITableView.automaticDimension
+        tableView.register(PostTableViewCell.self, forCellReuseIdentifier: PostTableViewCell.identifier )
+        return tableView
+    }()
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        self.title = "Мой профиль"
-//        tableView.delegate = self
-//        tableView.dataSource = self
-        setupProfileViewController()
+        layout()
     }
     
-    
+    private func layout() {
+        view.addSubview(tableView)
+        tableView.tableHeaderView = ProfileTableHeaderView(frame: CGRect(x: 0, y: 0, width: tableView.frame.width, height: 250))
+        tableView.backgroundColor = .systemGray4
+        
+        NSLayoutConstraint.activate([
+            tableView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor),
+            tableView.leadingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leadingAnchor),
+            tableView.trailingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.trailingAnchor),
+            tableView.bottomAnchor.constraint(equalTo: view.safeAreaLayoutGuide.bottomAnchor),
+        ])
+    }
 }
 
+
+extension ProfileViewController: UITableViewDataSource {
+    
+    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        return self.postModel.count
+    }
+    
+    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        let cell: PostTableViewCell = tableView.dequeueReusableCell(withIdentifier: PostTableViewCell.identifier, for: indexPath) as! PostTableViewCell
+        cell.setupCell(post: postModel[indexPath.row])
+        return cell
+    }
+}
+
+
+extension ProfileViewController: UITableViewDelegate {
+    
+    func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
+        return UITableView.automaticDimension
+    }
+    
+    func tableView(_ tableView: UITableView, canEditRowAt indexPath: IndexPath) -> Bool {
+        return true
+    }
+    
+    func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCell.EditingStyle, forRowAt indexPath: IndexPath) {
+        if editingStyle == .delete {
+            postModel.remove(at: indexPath.row)
+            tableView.deleteRows(at: [indexPath], with: .fade)
+        }
+    }
+}
+
+
+
+
+//protocol DetailDelegate: AnyObject {
+//    func changeText(_ text: String, indexPath: IndexPath)
+//}
+//
+//class ProfileViewController: UIViewController {
+//
+//    private let profileHeaderView = ProfileHeaderView()
+//
+//
+//    // добавляем элементы на контроллер
+//    func setupProfileViewController() {
+//        profileHeaderView.translatesAutoresizingMaskIntoConstraints = false
+//        view.addSubview(profileHeaderView)
+//
+//
+//        NSLayoutConstraint.activate([
+//            profileHeaderView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor),
+//            profileHeaderView.leadingAnchor.constraint(equalTo: view.leadingAnchor),
+//            profileHeaderView.trailingAnchor.constraint(equalTo: view.trailingAnchor),
+//            profileHeaderView.heightAnchor.constraint(equalToConstant: 220)
+//        ])
+//    }
+//
+//
+//    override func viewDidLoad() {
+//        super.viewDidLoad()
+//        self.title = "Мой профиль"
+////        tableView.delegate = self
+////        tableView.dataSource = self
+//        setupProfileViewController()
+//    }
+//
+//
+//}
+//
